move under python dir
