+++ release/create_release_pull_request.py
@@ -6,7 +6,7 @@ from os.path import abspath
 
 import semver
 
-from scripts.publish import assert_version_is_not_published
+from release.validate_version import assert_version_is_not_published
 
 SCRIPTS_DIR = os.path.dirname(abspath(getsourcefile(lambda: 0)))
 REPO_ROOT_DIR = os.path.dirname(SCRIPTS_DIR)
